+++ api/src/routers/services.ts
@@ -9,7 +9,7 @@ import { StreamsService } from '../services/streams-service';
 import { SubscriptionService } from '../services/subscription-service';
 import { UserService } from '../services/user-service';
 import { VerificationService } from '../services/verification-service';
-import { readRootIdentity } from '../setup/utilities';
+import { KeyResolver } from '../setup/key-resolver';
 import { Logger } from '../utils/logger';
 
 const { serverSecret, identityConfig, serverIdentityId, jwtExpiration, streamsConfig } = CONFIG;
@@ -34,7 +34,7 @@ export const verificationService = new VerificationService(
 	ssiService,
 	userService,
 	{
-		serverIdentityId: readRootIdentity(serverIdentityId),
+		serverIdentityId: new KeyResolver().resolve(serverIdentityId),
 		serverSecret,
 		keyCollectionSize: KEY_COLLECTION_SIZE
 	},
