+++ api/src/routes/verification/index.ts
@@ -10,13 +10,14 @@ import * as KeyCollectionLinksDb from '../../database/verifiable-credentials';
 import { AuthorizationService } from '../../services/authorization-service';
 import { VerifiableCredentialPersistence } from '../../models/types/key-collection';
 import { ILogger } from '../../utils/logger';
-import { readRootIdentity } from '../../setup/utilities';
 import * as _ from 'lodash';
+import { KeyResolver } from '../../setup/key-resolver';
 
 export class VerificationRoutes {
 	constructor(
 		private readonly verificationService: VerificationService,
 		private readonly authorizationService: AuthorizationService,
+		private readonly keyResolver: KeyResolver,
 		private readonly config: Config,
 		private readonly logger: ILogger
 	) {}
@@ -42,7 +43,7 @@ export class VerificationRoutes {
 
 			const vc: VerifiableCredentialJson = await this.verificationService.verifyIdentity(
 				subject,
-				readRootIdentity(this.config.serverIdentityId),
+				this.keyResolver.resolve(this.config.serverIdentityId),
 				initiatorVC?.credentialSubject?.id || requestUser.identityId
 			);
 
@@ -78,7 +79,7 @@ export class VerificationRoutes {
 				throw error;
 			}
 
-			await this.verificationService.revokeVerifiableCredential(vcp, readRootIdentity(this.config.serverIdentityId));
+			await this.verificationService.revokeVerifiableCredential(vcp, this.keyResolver.resolve(this.config.serverIdentityId));
 
 			res.sendStatus(StatusCodes.OK);
 		} catch (error) {
