+++ api/src/routes/verification/tests/bundled.test.ts
@@ -35,7 +35,7 @@ describe('test authentication routes', () => {
 			},
 			LoggerMock
 		);
-		verificationRoutes = new VerificationRoutes(verificationService, authorizationService, config, LoggerMock);
+		verificationRoutes = new VerificationRoutes(verificationService, authorizationService, { resolve: (x) => x }, config, LoggerMock);
 
 		res = {
 			send: sendMock,
