+++ api/src/routes/verification/tests/check-vc.test.ts
@@ -39,7 +39,7 @@ describe('test authentication routes', () => {
 			},
 			LoggerMock
 		);
-		verificationRoutes = new VerificationRoutes(verificationService, authorizationService, config, LoggerMock);
+		verificationRoutes = new VerificationRoutes(verificationService, authorizationService, { resolve: (x) => x }, config, LoggerMock);
 
 		res = {
 			send: sendMock,
