+++ api/src/routes/verification/tests/trusted-roots.test.ts
@@ -37,7 +37,7 @@ describe('test authentication routes', () => {
 			},
 			LoggerMock
 		);
-		verificationRoutes = new VerificationRoutes(verificationService, authorizationService, config, LoggerMock);
+		verificationRoutes = new VerificationRoutes(verificationService, authorizationService, { resolve: (x) => x }, config, LoggerMock);
 
 		res = {
 			send: sendMock,
