+++ api/src/routes/verification/tests/verify-user.test.ts
@@ -41,7 +41,8 @@ describe('test authentication routes', () => {
 			},
 			LoggerMock
 		);
-		verificationRoutes = new VerificationRoutes(verificationService, authorizationService, config, LoggerMock);
+
+		verificationRoutes = new VerificationRoutes(verificationService, authorizationService, { resolve: (x) => x }, config, LoggerMock);
 
 		res = {
 			send: sendMock,
